Fix issue with shop loading unnecessarily
The caching enhancements seemed to irritate an existing bug that caused
the shop data to be submitted unnecessarily.  Due to ordering issues
around how the tab data was returned, or if users possibly moved items,
the shop data hash would change causing an incorrect refresh of the
shop.

This  will probably fix issue #223, as if user forced refresh in short
amount of time and made no changes this error would result.