Add 100ms delay to updating item info to speed up multi-select
Currently when multi-selecting the items are forced to update after each
selection so when you select a large range of items it's slow and clumsy
as items must refresh as you pass over each item.

This update adds a 100ms delay to updating item info which is reset when
ever a new item is added to the selection.  So when quickly selecting a
range of items only the last will be shown.